Initial Menu Display (#38)
Defines Menu and Menu item types to represent a menu and the items it contains, adds a MenuService interface and a FixedMenuService implementation that gives back a static menu whenever todaysMenu or menuFor.

Also adds a MenuPanel component to hold the lists of items on the menu and a MenuSection component to represent each list of menu items (e.g. main courses or desserts).

Finally, creates an instance of FixedMenuService in main.ts and provides it to App.Vue as a prop.